feat: add support to replace model in resource stub
