+++ src/Console/Concerns/ReplacesModel.php
@@ -1,0 +1,58 @@
+<?php
+/*
+ * Copyright 2024 Cloud Creativity Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+declare(strict_types=1);
+
+namespace LaravelJsonApi\Laravel\Console\Concerns;
+
+use Illuminate\Support\Str;
+use function array_keys;
+use function array_values;
+use function str_replace;
+
+trait ReplacesModel
+{
+
+    /**
+     * @param string $stub
+     * @param string $model
+     * @return string
+     */
+    protected function replaceModel(string $stub, string $model): string
+    {
+        $model = str_replace('/', '\\', $model);
+
+        if (Str::startsWith($model, '\\')) {
+            $namespacedModel = trim($model, '\\');
+        } else {
+            $namespacedModel = $this->qualifyModel($model);
+        }
+
+        $model = class_basename($model);
+
+        $replace = [
+            '{{ namespacedModel }}' => $namespacedModel,
+            '{{namespacedModel}}' => $namespacedModel,
+            '{{ model }}' => $model,
+            '{{model}}' => $model,
+        ];
+
+        return str_replace(
+            array_keys($replace), array_values($replace), $stub
+        );
+    }
+}
